Kotlin and version update (0.7.10)
