+++ jipp-core/build.gradle
@@ -55,6 +55,10 @@ test {
     finalizedBy jacocoTestReport
 }
 
+jacoco {
+    toolVersion = "0.8.7"
+}
+
 jacocoTestReport {
     reports {
         xml.enabled = true // coveralls plugin depends on xml format report
