+++ jipp-pdl/src/test/kotlin/util/KotlinTest.kt
@@ -1,6 +1,8 @@
 // Copyright 2018 HP Development Company, L.P.
 // SPDX-License-Identifier: MIT
 
+package util
+
 import org.junit.Assert
 
 object KotlinTest {
